pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "netwatch"
version = "0.1.0"
description = "Local ISP speed logger with AI-powered insights via Groq"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["isp", "speedtest", "network", "ai", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Monitoring",
]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "groq>=0.4.0",
    "python-dotenv>=1.0",
    "speedtest-cli>=2.1.3",
    "schedule>=1.2.0",
    "plotext>=5.2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
]

[project.scripts]
netwatch = "netwatch.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/netwatch"
Issues = "https://github.com/yourusername/netwatch/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.setuptools.packages.find]
where = ["."]
include = ["netwatch*"]
